# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation untuk API Manajemen Senjata",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/senjata": {
            "get": {
                "tags": [
                    "Senjata"
                ],
                "summary": "Menampilkan semua senjata",
                "operationId": "8dbf22610d7ecef53e9faf64a09d4185",
                "responses": {
                    "200": {
                        "description": "Data senjata ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Senjata"
                ],
                "summary": "Menyimpan senjata baru",
                "operationId": "ece9e9afadab98cabeb5a4fdf5867ad9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_senjata",
                                    "id_jenis",
                                    "id_gudang",
                                    "stok",
                                    "kaliber",
                                    "nomor_seri"
                                ],
                                "properties": {
                                    "nama_senjata": {
                                        "type": "string"
                                    },
                                    "id_jenis": {
                                        "type": "integer"
                                    },
                                    "id_gudang": {
                                        "type": "integer"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    },
                                    "kaliber": {
                                        "type": "string"
                                    },
                                    "nomor_seri": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Senjata berhasil dibuat"
                    },
                    "500": {
                        "description": "Terjadi kesalahan saat menyimpan data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/senjata/{id}": {
            "get": {
                "tags": [
                    "Senjata"
                ],
                "summary": "Menampilkan senjata berdasarkan ID",
                "operationId": "7e65da47b172b332f0d6954c5db4a2f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Senjata berhasil ditemukan"
                    },
                    "404": {
                        "description": "Senjata tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Senjata"
                ],
                "summary": "Mengupdate senjata berdasarkan ID",
                "operationId": "7ef6e51095b55c6c5664553eb9873c9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_senjata",
                                    "id_jenis",
                                    "id_gudang",
                                    "stok",
                                    "kaliber",
                                    "nomor_seri"
                                ],
                                "properties": {
                                    "nama_senjata": {
                                        "type": "string"
                                    },
                                    "id_jenis": {
                                        "type": "integer"
                                    },
                                    "id_gudang": {
                                        "type": "integer"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    },
                                    "kaliber": {
                                        "type": "string"
                                    },
                                    "nomor_seri": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senjata berhasil diupdate"
                    },
                    "404": {
                        "description": "Senjata tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Senjata"
                ],
                "summary": "Menghapus senjata berdasarkan ID",
                "operationId": "15bcab3fa91f1c40f9ea568d22d2d72d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Senjata berhasil dihapus"
                    },
                    "404": {
                        "description": "Senjata tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "sanctum": []
        }
    }
}